Add files via upload
This code was developed to correct a problem which occured when attempting to measure the interval between data samples tansmitted via both UDP and RS485. It was found that an accurate time could not be determined using mean or average due to excessive jitter on the time measurements. This was for Linux  clock_gettime(CLOCK_REALTIME and Windows QueryPerformanceCounter. However is was discovered that an accurate determination could be made by calculating a mode ( most common) value.